application\models\Customer_Model.php application\modules\Customer\controllers\Customer.php application\modules\Customer\views\Edit_customer_info.php application\modules\Customer\views\client_user_list.php application\modules\Customer\views\customer_business_location_add.php application\modules\Customer\views\customer_business_location_edit.php application\modules\Customer\views\customer_business_location_list.php

+++ application/modules/Customer/views/client_user_list.php
@@ -93,7 +93,7 @@
                       <i class="fa fa-pencil blue" data-toggle="tooltip" data-placement="top" title="" data-original-title="Edit">
                       </i>
                     </a>
-                    <a href="<?php echo base_url()?>Customer/delete_client_user?client_id=<?php echo $clientinfo->id; ?>" title="Delete">
+                    <a href="<?php echo base_url()?>Customer/delete_client_user?client_id=<?php echo $clientinfo->id; ?>" title="Delete" onclick="return confirm('Are you sure you want to delete this item?');">
                       <i class="fa fa-trash red" data-toggle="tooltip" data-placement="top" title="" data-original-title="Delete">
                       </i> 
                     </a>
