+++ application/modules/Customer/views/customer_business_location_list.php
@@ -25,11 +25,11 @@
             <div class="row clearfix">
               <div class="col-lg-12 col-md-12 col-sm-12 col-xs-12">
                 <ul class="flex-container flex-container-head nowrap">
-                  <li>Sr. No.
+                  <li class="flex-item">Sr. No.
                   </li>
-                  <li class="">Location
+                  <li class="flex-item">Location
                   </li>
-                  <li class="">Address
+                  <li class="flex-item">Address
                   </li>
                   <li class="flex-item">Contact Person
                   </li>
@@ -56,21 +56,21 @@
             <div class="row clearfix">
               <div class="col-lg-12 col-md-12 col-sm-12 col-xs-12">
                 <ul class="flex-container nowrap">
-                  <li class=""><?php echo $location->id; ?>
+                  <li class="flex-item"><?php echo $location->id; ?>
                   </li>
-                  <li class="wrap-word"><?php echo $location->location_name; ?>
+                  <li class="flex-item"><?php echo $location->location_name; ?>
                   </li>
-                  <li class="wrap-word"><?php echo $location->address; ?>
+                  <li class="flex-item"><?php echo $location->address; ?>
                   </li>
                   <li class="flex-item"><?php echo $location->contact_person_name; ?>
                   </li>
-                  <li class="flex-item"><?php echo $location->city; ?>
+                  <li class="flex-item"><?php echo $location->city_name; ?>
                   </li>
-                  <li class="flex-item"><?php echo $location->state; ?>
+                  <li class="flex-item"><?php echo $location->state_name; ?>
                   </li>
                   <li class="flex-item"><?php echo $location->pincode; ?>
                   </li>
-                  <li class="flex-item"><?php echo $location->country; ?>
+                  <li class="flex-item"><?php echo $location->country_name; ?>
                   </li>
                   <li class="flex-item"><?php echo $location->telephone; ?>
                   </li>
@@ -84,7 +84,7 @@
                       <i class="fa fa-pencil blue" data-toggle="tooltip" data-placement="top" title="" data-original-title="Edit">
                       </i>
                     </a>
-                    <a href="<?php echo base_url()?>Customer/delete_business_location?business_id=<?php echo $location->id; ?>" title="Delete">
+                    <a href="<?php echo base_url()?>Customer/delete_business_location?business_id=<?php echo $location->id; ?>"  title="Delete" onclick="return confirm('Are you sure you want to delete this item?');">
                       <i class="fa fa-trash red" data-toggle="tooltip" data-placement="top" title="" data-original-title="Delete">
                       </i> 
                     </a>
